# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Lorem Ipsum",
        "contact": {
            "email": "rifkyr.personal@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/showuser": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Show Data Users",
                "description": "Get All Data User",
                "operationId": "show user",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/createaccount": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Create Account & Generate Token For Getting Address Book Data",
                "description": "Please Make sure you save or remember the token after the token is generated ",
                "operationId": "createaccount",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string||email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Testing",
                                    "email": "Testing@testing.com",
                                    "password": "testing"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/deleteaccount": {
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Delete Account",
                "description": "Delete Data User",
                "operationId": "deleteaccount",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Delete Account User",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/store": {
            "post": {
                "tags": [
                    "Address Book"
                ],
                "summary": "Store Address Book",
                "description": "Insert Data Address Book",
                "operationId": "store adressbook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string||email"
                                    },
                                    "phone_number": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Testing",
                                    "email": "Testing@testing.com",
                                    "phone_number": "testing",
                                    "address": "testing"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/show": {
            "get": {
                "tags": [
                    "Address Book"
                ],
                "summary": "Show Address Book",
                "description": "Get All Data Address Book",
                "operationId": "show adressbook",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    [
                                        {
                                            "id": "1",
                                            "name": "Testing",
                                            "email": "testing@testing.com",
                                            "phone_number": "234456678",
                                            "address": "Testing Street No 14"
                                        }
                                    ]
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/update": {
            "put": {
                "tags": [
                    "Address Book"
                ],
                "summary": "Update Address Book",
                "description": "Update Data Address Book",
                "operationId": "update adressbook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Add Id Address Book",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string||email"
                                    },
                                    "phone_number": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Testing",
                                    "email": "Testing@testing.com",
                                    "phone_number": "testing",
                                    "address": "testing"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/delete": {
            "delete": {
                "tags": [
                    "Address Book"
                ],
                "summary": "Delete Address Book",
                "description": "Delete Data Address Books",
                "operationId": "delete addressbook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Add Id Address Book",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}